Collapse event metadata.
