+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/consensus/RoundElections.java
@@ -21,7 +21,6 @@ import static com.swirlds.logging.legacy.LogMarker.CONSENSUS_VOTING;
 import com.swirlds.common.platform.NodeId;
 import com.swirlds.common.utility.IntReference;
 import com.swirlds.platform.Utilities;
-import com.swirlds.platform.event.EventMetadata;
 import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.state.MinGenInfo;
 import com.swirlds.platform.system.events.EventConstants;
@@ -149,7 +148,7 @@ public class RoundElections {
                 .mapToLong(EventImpl::getGeneration)
                 .min()
                 .orElse(EventConstants.GENERATION_UNDEFINED);
-        allJudges.forEach(EventMetadata::setJudgeTrue);
+        allJudges.forEach(EventImpl::setJudgeTrue);
 
         return allJudges;
     }
